+++ documentation/HISTORY.md
@@ -11,7 +11,10 @@
 - Function types into sample model User corrected;
 - Configuration directory moved to root;
 - Library directory moved to /springy;
-- Vendor directory moved to root.
+- Vendor directory moved to root;
+- Kernel received the responsability to find controller and start the application;
+- Moved autoload and error handlers initiations to helper script;
+- Starter script simplified;.
 
 ## Version 3
 ### 3.6.3